requirements_dev.txt
-r requirements_test.txt
-r requirements.txt

django-extensions
invoke
twine
wheel
zest.releaser
